% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{analyse_change}
\alias{analyse_change}
\title{Get a set of figures and tables for analysing the change in the variable of interest}
\usage{
analyse_change(
  data,
  id,
  x,
  y,
  rain_side = NULL,
  nudge_y = NULL,
  color_fill = NULL,
  color_stat = NULL,
  labs_1x = NULL,
  labs_1y = NULL,
  labs_2x = NULL,
  labs_2y = NULL,
  labs_3x = NULL,
  labs_3y = NULL,
  labs_4x = NULL,
  labs_4y = NULL,
  labs_5x = NULL,
  labs_5y = NULL,
  labs_6x = NULL,
  labs_6y = NULL
)
}
\arguments{
\item{data}{A dataframe.}

\item{id}{A character value to indicate the name of the variable designating the
identities of the observations.}

\item{x}{A character value to indicate the name of the variable to be plotted on the X axis.}

\item{y}{A character value to indicate the name of the variable to be plotted on the Y axis.}

\item{rain_side}{A character value to indicate how you want the rainclouds displayed,
right ("r"), left ("l") or flanking ("f"), for a 1-by-1 flanking raincloud use ("f1x1")
and for a 2-by-2 use ("f2x2").}

\item{nudge_y}{A numeric value to vertically move the labels of the graphic showing the shift function.}

\item{color_fill}{A character value to set the filling color.}

\item{color_stat}{A character value to set the color of the statistical summaries.}

\item{labs_1x}{A character string to name the X axis of the 1rst graphic (marginal distributions).}

\item{labs_1y}{A character string to name the Y axis of the 1rst graphic (marginal distributions).}

\item{labs_2x}{A character string to name the X axis of the 2nd graphic (pairwise differences).}

\item{labs_2y}{A character string to name the Y axis of the 2nd graphic (pairwise differences).}

\item{labs_3x}{A character string to name the X axis of the 3rd graphic (bivariate relationship).}

\item{labs_3y}{A character string to name the Y axis of the 3rd graphic (bivariate relationship).}

\item{labs_4x}{A character string to name the X axis of the 4th graphic (quantile shifts).}

\item{labs_4y}{A character string to name the Y axis of the 4th graphic (quantile shifts).}

\item{labs_5x}{A character string to name the X axis of the 5th graphic (shift function).}

\item{labs_5y}{A character string to name the Y axis of the 5th graphic (shift function).}

\item{labs_6x}{A character string to name the X axis of the 6th graphic (difference asymmetry function).}

\item{labs_6y}{A character string to name the Y axis of the 6th graphic (difference asymmetry function).}
}
\value{
A ggplot object.
}
\description{
Get a set of figures and tables for analysing the change in the variable of interest
}
