% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_distribution.R
\name{analyse_distribution}
\alias{analyse_distribution}
\title{Plot distribution}
\usage{
analyse_distribution(data, var)
}
\arguments{
\item{data}{A dataframe.}

\item{var}{A character value to indicate the name of a numeric variable to be plotted.}
}
\value{
A ggplot object.
}
\description{
Function to graphically analyse the distribution of a continuous variable
using a raincloud plot and a qq-plot
}
