% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BARRIERS}
\alias{BARRIERS}
\title{Barriers to physical activity}
\format{
\subsection{\code{BARRIERS}}{

A data frame with 77 rows and 12 columns:
\describe{
\item{patient}{Participant's ID.}
\item{num_visit}{1 = visit at the end of the cardiac rehabilitation program; 2 = visit at 6 months post program; 3 = visit at 12 months post program.}
\item{trop_vieux}{"Too old"; 1 = "yes"; 0 = "no".}
\item{manque_interet}{"Lack of interest"; 1 = "yes"; 0 = "no".}
\item{effort_import_fatig}{"Heavy effort / too tired"; 1 = "yes"; 0 = "no".}
\item{manque_temps}{"Lack of time"; 1 = "yes"; 0 = "no".}
\item{meteo_defavorable}{"Unfavourable weather"; 1 = "yes"; 0 = "no".}
\item{deplacements_diff}{"Difficulty to move"; 1 = "yes"; 0 = "no".}
\item{cout_trop_eleve}{"Too costly"; 1 = "yes"; 0 = "no".}
\item{crainte_blessures_douleurs}{"Fear of injury / pain"; 1 = "yes"; 0 = "no".}
\item{isolement_faible_RS }{"Social isolation / weak social network"; 1 = "yes"; 0 = "no".}
\item{autres}{"Other reasons".}
\item{autre_precision_obstacle}{"Other details about barriers".}
}
}
}
\source{
APA & Co project
}
\usage{
BARRIERS
}
\description{
Participants responses to the questionnaire assessing barriers to physical activity
12 months after the end of the cardiac rehabilitation program.
}
\keyword{datasets}
