% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_t_tost}
\alias{do_t_tost}
\title{Perform TOSTs for several equivalence bounds}
\usage{
do_t_tost(data, var, group, ref)
}
\arguments{
\item{data}{A dataframe.}

\item{var}{A character value to indicate the name of the variable to be plotted.}

\item{group}{A character value to indcate the name of the variable relating to the tested conditions.}

\item{ref}{A character value to indicate for which condition the reference mean shoud be computed.}
}
\value{
A table of results.
}
\description{
Perform TOSTs for several equivalence bounds
}
