% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMAPS}
\alias{EMAPS}
\title{EMAPS}
\format{
\subsection{\code{EMAPS}}{

A data frame with 241 rows and 20 columns:
\describe{
\item{patient}{Participant's ID.}
\item{num_visit}{1 = visit at the end of the cardiac rehabilitation program; 2 = visit at 6 months post program; 3 = visit at 12 months post program.}
\item{AP_q1 to AP_q18}{Answers to the EMAPS items.}
}
}
}
\source{
APA & Co project
}
\usage{
EMAPS
}
\description{
Participants responses to the EMAPS at 0, 6 and 12 months after the
end of the cardiac rehabilitation program.
}
\keyword{datasets}
