% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DB_6MWT.R
\name{get_DB_6MWT}
\alias{get_DB_6MWT}
\title{Get final 6MWT database}
\usage{
get_DB_6MWT(data1, data2, data3)
}
\arguments{
\item{data1}{'INCLUSION_cleaned' data frame (from APA&Co project analytical pipeline).}

\item{data2}{'VISIT_6M_cleaned' data frame (from APA&Co project analytical pipeline).}

\item{data3}{'VISIT_12M_cleaned' data frame (from APA&Co project analytical pipeline).}
}
\value{
A data frame.
}
\description{
Get final 6MWT database
}
