% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DB_EMAPS.R
\name{get_DB_EMAPS}
\alias{get_DB_EMAPS}
\title{Get final EMAPS database (APA & Co project)}
\usage{
get_DB_EMAPS(data)
}
\arguments{
\item{data}{'EMAPS' dataframe (from APA & Co project analysis pipeline).}
}
\value{
A dataframe.
}
\description{
Get final EMAPS database (APA & Co project)
}
