% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DB_IPAQ.R
\name{get_DB_IPAQ}
\alias{get_DB_IPAQ}
\title{Get final IPAQ database}
\usage{
get_DB_IPAQ(data)
}
\arguments{
\item{data}{'IPAQ' data frame (from APA & Co project analytical pipeline).}
}
\value{
A data frame.
}
\description{
Get final IPAQ database
}
