% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DB_IPAQ_6_12.R
\name{get_DB_IPAQ_6_12}
\alias{get_DB_IPAQ_6_12}
\title{Filtering DB_IPAQ: Keep the rows for months 6 and 12, and keep the participants with data
at both 6 and 12 months (APA & Co project)}
\usage{
get_DB_IPAQ_6_12(data)
}
\arguments{
\item{data}{'DB_IPAQ' dataframe (from APA & Co project analysis pipeline).}
}
\value{
A dataframe.
}
\description{
Filtering DB_IPAQ: Keep the rows for months 6 and 12, and keep the participants with data
at both 6 and 12 months (APA & Co project)
}
