% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_barriers.R
\name{get_plot_barriers}
\alias{get_plot_barriers}
\title{Get the plot showing the results relating to the barriers to physical activity (APA & Co project)}
\usage{
get_plot_barriers(data)
}
\arguments{
\item{data}{'BARRIERS_cleaned' dataframe (from APA & Co project analysis pipeline).}
}
\value{
A ggplot object.
}
\description{
Get the plot showing the results relating to the barriers to physical activity (APA & Co project)
}
