% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_IPAQ_all.R
\name{get_plot_IPAQ_all}
\alias{get_plot_IPAQ_all}
\title{Get plot for all IPAQ MET-min/week data}
\usage{
get_plot_IPAQ_all(data)
}
\arguments{
\item{data}{'IPAQ' data frame (from APA & Co project analytical pipeline).}
}
\value{
A ggplot object.
}
\description{
Get plot for all IPAQ MET-min/week data
}
