% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%+\%}
\alias{\%+\%}
\title{Define a custom + operator allowing to handle NAs}
\usage{
x \%+\% y
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
A numeric value.
}
\description{
Define a custom + operator allowing to handle NAs
}
