% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INCLUSION}
\alias{INCLUSION}
\title{Physical, clinical, and functional characteristics at the inclusion visit}
\format{
\subsection{\code{INCLUSION}}{

A data frame with 83 rows and 8 columns:
\describe{
\item{patient}{Participant's ID.}
\item{age}{Age in years.}
\item{sex}{1 = male; 2 = female.}
\item{height}{Height in cm.}
\item{weight}{Weight in kg.}
\item{angioplasty}{1 = Has had an angioplasty; 0 = Has not had an angioplasty.}
\item{bypas}{1 = Has had a bypass surgery; 0 = Has not had a bypass surgery.}
\item{DIST_6WT_M0}{6-min walking test distance in m.}
}
}
}
\source{
APA&Co project
}
\usage{
INCLUSION
}
\description{
Participants data relating to the physical, clinical, and functional profiles
at the end of the cardiac rehabilitation program.
}
\keyword{datasets}
