% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read dataset from a package}
\usage{
read_data(data_name, package_name)
}
\arguments{
\item{data_name}{A character value naming the dataset to be loaded.}

\item{package_name}{A character value naming the package from which data should be loaded.}
}
\value{
A dataset.
}
\description{
Read dataset from a package
}
