% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_figure.R
\name{save_figure}
\alias{save_figure}
\title{Save figure}
\usage{
save_figure(
  filename,
  ...,
  scaling = 0.4,
  height = 16,
  width = 16,
  units = "cm",
  res = 400
)
}
\arguments{
\item{filename}{Cf. ?ragg::agg_tiff.}

\item{...}{The object to be plotted.}

\item{scaling}{Cf. ?ragg::agg_tiff.}

\item{height}{Cf. ?ragg::agg_tiff.}

\item{width}{Cf. ?ragg::agg_tiff.}

\item{units}{Cf. ?ragg::agg_tiff.}

\item{res}{Cf. ?ragg::agg_tiff.}
}
\value{
A character string indicating the path to the saved figure.
}
\description{
Save figure
}
