% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_rainclouds}
\alias{view_rainclouds}
\title{Make a polished raincloud plot for visualising the change}
\usage{
view_rainclouds(
  data,
  id,
  x,
  y,
  rain_side = NULL,
  color_fill = NULL,
  color_stat = NULL,
  labs_x = NULL,
  labs_y = NULL
)
}
\arguments{
\item{data}{A dataframe.}

\item{id}{A character value to indicate the name of the variable designating the
identities of the observations.}

\item{x}{A character value to indicate the name of the variable to be plotted on the X axis.}

\item{y}{A character value to indicate the name of the variable to be plotted on the Y axis.}

\item{rain_side}{A character value to indicate how you want the rainclouds displayed,
right ("r"), left ("l") or flanking ("f"), for a 1-by-1 flanking raincloud use ("f1x1")
and for a 2-by-2 use ("f2x2").}

\item{color_fill}{A character value to set the filling color.}

\item{color_stat}{A character value to set the color of the statistical summaries.}

\item{labs_x}{A character string to name the X axis.}

\item{labs_y}{A character string to name the Y axis.}
}
\value{
A ggplot object.
}
\description{
Make a polished raincloud plot for visualising the change
}
