% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VISIT_12M}
\alias{VISIT_12M}
\title{Six-minute walking test performance at 12 months}
\format{
\subsection{\code{VISIT_12M}}{

A data frame with 77 rows and 3 columns:
\describe{
\item{patient}{Participant's ID.}
\item{weight}{Weight in kg.}
\item{DIST_6WT_M6}{6-min walking test distance in m.}
}
}
}
\source{
APA&Co project
}
\usage{
VISIT_12M
}
\description{
Participants data relating to the 6-min walking test 12 months after
the end of the cardiac rehabilitation program.
}
\keyword{datasets}
