% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_change_IPAQ.R
\name{analyse_change_IPAQ}
\alias{analyse_change_IPAQ}
\title{Get results for the change in IPAQ MET-min/week between 6 and 12 months}
\usage{
analyse_change_IPAQ(data)
}
\arguments{
\item{data}{'DB_IPAQ_6_12' data frame (from APA & Co project analytical pipeline).}
}
\value{
A list with one figure and two tables (one for the shift function, and one for
the difference asymmetry function).
}
\description{
Get results for the change in IPAQ MET-min/week between 6 and 12 months
}
