% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DB_6MWT_0_12.R
\name{get_DB_6MWT_0_12}
\alias{get_DB_6MWT_0_12}
\title{Filtering DB_6MWT}
\usage{
get_DB_6MWT_0_12(data)
}
\arguments{
\item{data}{'DB_6MWT' data frame (from APA&Co project analytical pipeline).}
}
\value{
A data frame.
}
\description{
Keep the rows for months 0 and 12, and keep the participants with data
at both 0 and 12 months.
}
