% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DB_EMAPS.R
\name{get_DB_EMAPS}
\alias{get_DB_EMAPS}
\title{Get final EMAPS database}
\usage{
get_DB_EMAPS(data)
}
\arguments{
\item{data}{'EMAPS' data frame (from APA&Co project analytical pipeline).}
}
\value{
A data frame.
}
\description{
Get final EMAPS database
}
