% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_6MWT_all.R
\name{get_plot_6MWT_all}
\alias{get_plot_6MWT_all}
\title{Get plot for all 6MWT distance data}
\usage{
get_plot_6MWT_all(data)
}
\arguments{
\item{data}{'DB_6MWT' data frame (from APA&Co project analytical pipeline).}
}
\value{
A ggplot object.
}
\description{
Get plot for all 6MWT distance data
}
