% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_barriers.R
\name{get_plot_BARRIERS}
\alias{get_plot_BARRIERS}
\title{Get the plot showing the results relating to the barriers to physical activity
at 12 months}
\usage{
get_plot_BARRIERS(data)
}
\arguments{
\item{data}{'BARRIERS_cleaned' data frame (from APA&Co project analytical pipeline).}
}
\value{
A ggplot object.
}
\description{
Get the plot showing the results relating to the barriers to physical activity
at 12 months
}
