% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_EMAPS_all.R
\name{get_plot_EMAPS_all}
\alias{get_plot_EMAPS_all}
\title{Get plot for all EMAPS scores data}
\usage{
get_plot_EMAPS_all(data)
}
\arguments{
\item{data}{'DB_EMAPS' data frame (from APA&Co project analytical pipeline).}
}
\value{
A ggplot object.
}
\description{
Get plot for all EMAPS scores data
}
